Add method for saving data as csv
The method for saving data as csv was pretty straightforward
(just throwing the data into pandas and calling .to_csv), but
this required figuring out a way to encode non-ascii characters.
I chose to use .encode('ascii', 'xmlcharrefreplace') because
it seemed easy enough to use, and because I found a reasonably
easy way to decode the values (using HTMLParser.unescape()).