Rework general_utilities to add parsing flexibility
I wanted the general_utilities functions to have the option to parse not
just text from html tags, but other attributes/tags as well (i.e. 'a'
tags, 'href' attributes, etc.). This is largely driven by wanting to be
able to grab these in parsing another page (commit about to come), but
required reworking some of the album_of_the_year.py to account for the
added flexibility in the general_utilities.py.