Initialize end_year_album_lists.py
This initial commit sets up the beginning stages of grabbing the end of
year album lists from different critic lists. Currently, it just grabs
the hrefs so I will be able to go to those pages and grab and parse
their lists. It makes good use of the previously added flexibility to
the general_utilities.py (commit prior to this).