Initial commit for grabbing individual album info.
This is the base shell script that I will be operating off of for
grabbing invidual info for each album (specifically user and critic
rating info.). There are currently two functions in it that are written
but need to be finished: find_user_score and find_critic_score. Right
now they just have 'pass' in them.