Update albums_of_year_lst_ind.py to output to mongo
This involved editing the output_to_mongo function in
general_utilities.py to handle the case that we wanted to update an
existing observation in mongo (and not just insert new observations).
This edit involved just building another function that would update
observations in mongo by a particular key that was present for that
observation and specified by the user.