Rework album_of_the_year code with general utilities
Some of the code in the album_of_the_year scraper was still using
it's own general_utilities.py file, and as such there was some code that
was written twice in two different places. For that code, it was taken
out of the album_of_the_year files and the paths were instead altered so
that it could use the code from the general_utilities folder. For the
code that wasn't written in two places but was still kind of a general
utility kind of thing, it was left in the general_utlities.py file,
which was renamed to special_utilities.py.