Move code into RequestThreading object
For some reason, I didn't think about the fact that I should thread this
when I set it up. I thread all of the other job scrapers that I
built, but in having to work with Selenium here, I started off not
threading it. That was a mistake, and I can just as easily thread this
scraper like I did other ones.