Refactor code to make threading work across all jobs
In previous iterations, I was using the xpath search to try to identify
certain parts of a job posting, and misunderstood how it worked. As a
result, I was simply iterating over the same job posting over and over
again on a given page, rather than iterating over all job postings.

Searching by the xpath if you don't search for all elements that meet
that citeria simply gives back the first one that meets that xpath. So,
in reality I was just searching over the same job over and over again,
instead of searching over all jobs. This should fix that.