Alter code to use a get request for job search
Previously, I thought that I needed to actually use Selenium to click
the search boxes, and put input in them to search for jobs. I didn't
realize that selenium had a `.get()` method through which I could just
input the original query. I think this makes more sense, and it has the
perk that I don't also have to search and click for a job radius or to
order by date (I can add these into the URL).