Move parsing of jobs and pages into general utility
Each one of the job scrapers that I built was using largely the same
code in order to parse the total number of jobs that resulted from a
query. One was even using it to find the total number of pages there
were to search over. Rather than have that in the 6 job scrapers that I
ave built (currently), in along with a DRY methodology it's probably
best to now have that function located in one place, and just a little
bit more general.