Fill in code to cycle through pages
Here, I've finished up and added one or two things necessary to
actually cycle through each page on a glassdoor search. This might not
end up working, as sometimes the pages freeze (could be me working too
much of my computer right now). Occasionally, it also stops you and asks
if you're a bot, which could also lead this not to work. However, the way
that the code is written up now, it would work barring these issues
(I'll have to come back to figure out if there's a way around the other
issues).