Move href threading code into a general utility
In coding up the CareerBuilder scraper, I realized that the threading I
was doing in order to query each job's href was exactly the same as the
threading I was doing in the Monster scraper. I put off moving it all
into a general utility because I wanted to do it all at once (i.e. for
the Monster and Careerbuilder scraper), and not have it happen in the
middle of developing the Careerbuilder scraper.