Update careerbuilder and monster nav utilities to fix bug
With the way that the careerbuilder and monster scrapers were built,
they were throwing errors in a number of different circumstances. The
first was because selenium was trying to access parts of the page before
they were actually rendered, and the second is just the odd case where
something goes wrong with the Internet. The code added in this commit
accounts for those two issues.

The first issue was accounted for just by adding in an implicit wait
time on the Selenium browser, whereas the second involved the inclusion
of a try/except block (not ideal, but necessary to keep the program
running given the error).