Fix missing element bug in monster and careerbuilder
For some odd reason, occasionally the monster and careerbuilder scrapers
lose elements (i.e. the error reads that they are not in the cash). I
hate throwing try/excepts in there like this, but I've tried a number of
other different things - moving code around, throwing some implicit
waits in there (which seemed to help a little), and even some explicit
time.sleeps (which didn't really help). I think this is really the only
option left that I see as feasible.